+++ src/SharePoint.Support.fs
@@ -3,17 +3,27 @@ open Fable.Import.SharePoint.SP
 open Fable.Core
 open Fable.Core.JsInterop
 open Fable.Import.Browser
+open Fable.PowerPack
 
 open Browser.Support
 
 open Microsoft.FSharp.Reflection
 
+[<Emit("log4javascript.getLogger().debug($0)")>]
+let log (message:string) : unit = jsNative
+
+[<Emit("log4javascript.getLogger().debug($0)")>]
+let logO (value:obj) : unit = jsNative
+
+[<Emit("log4javascript.getLogger().debug($0, $1)")>]
+let logO2 (message:string) (value:obj) : unit = jsNative
+
 let onQueryFailed sender (args:ClientRequestFailedEventArgs) =
     let message = sprintf "Request failed. %s \n%s\n " (args.get_message()) (args.get_stackTrace())
-    //failwith message
     log message
     logO args
     log "------------------------------------------------------------------------"
+    failwith message
 
 let nothingOnQueryFailed sender (args:ClientRequestFailedEventArgs) =
     ()
@@ -116,41 +126,27 @@ type ListDefinition = {
     Fields : FieldDefinition array
 }
 
-let executeQueryAsyncWithFallback (clientContext:ClientContext) (fallback) =
-    Async.FromContinuations( fun( cont, econt, ccont ) ->
-        clientContext.executeQueryAsync(
-            System.Func<_,_,_>(fun _ _ -> cont() ),
-            System.Func<_,_,_>(fallback)
-        )        
-    )
+[<Emit("jQuery.Deferred()")>]
+let deferred() = jsNative
 
-let executeQueryAsyncWithSuccessAndFallback (clientContext:ClientContext) (success) (fallback) =
-  Async.FromContinuations( fun( cont, econt, ccont ) ->
-      clientContext.executeQueryAsync(
-          System.Func<_,_,_>(success),
-          System.Func<_,_,_>(fallback)
-      )        
-  )
-
-let executeQueryAsyncWithSuccessCallback (clientContext:ClientContext) (onSuccess) =
-    executeQueryAsyncWithSuccessAndFallback clientContext onSuccess onQueryFailed
-
-let executeQueryAsync (clientContext:ClientContext) =
-    executeQueryAsyncWithFallback clientContext onQueryFailed
-
-let executeSilentQueryAsync (clientContext:ClientContext) =
-    executeQueryAsyncWithFallback clientContext nothingOnQueryFailed
-
+let executeQuery (clientContext:ClientContext) : Fable.Import.JS.Promise<unit> =
+    let deferred = deferred()
+    clientContext.executeQueryAsync(
+        System.Func<_,_,_>( fun _ arguments -> deferred?resolve(arguments) |> ignore ),
+        System.Func<_,_,_>( fun _ arguments -> deferred?reject(arguments) |> ignore )
+    )            
+    downcast deferred?promise()
 
 let createCustomList title url (clientContext : ClientContext) =
-    async {
+    promise {
         let web = clientContext.get_web()
         let listCollection = web.get_lists()
         clientContext.load(listCollection)
         let list1 = listCollection.getByTitle(title)
         clientContext.load(list1)
+        let execQuery () = clientContext |> executeQuery
         let doCreateList () = 
-                async {
+                promise {
                     let listCreationInfo = ListCreationInformation()
                     listCreationInfo.set_title(title)
                     listCreationInfo.set_url("Lists/"+url)
@@ -158,10 +154,11 @@ let createCustomList title url (clientContext : ClientContext) =
                     let list1 = web.get_lists().add(listCreationInfo)
 
                     clientContext.load(list1);
-                    do! executeQueryAsync clientContext
-                } |> Async.StartImmediate
-        do! executeQueryAsyncWithFallback clientContext  (  fun _ _ -> doCreateList () )                
-    } |> Async.StartImmediate
+                    do! execQuery ()
+                } 
+        do! execQuery ()
+        do! doCreateList ()
+    } |> Promise.iter(ignore)
 
 let private createCustomListInt title url (createContentType:bool) continue1 (listCollection:ListCollection) (web:Web) (clientContext : ClientContext) =
     let list1 = listCollection.getByTitle(title)
@@ -203,23 +200,15 @@ let private createCustomListInt title url (createContentType:bool) continue1 (li
     )
 
 let getListIdAsync title (listCollection:ListCollection) (clientContext : ClientContext) = 
-    async {
+    promise {
         let list1 = listCollection.getByTitle(title)
         clientContext.load(list1, "Id")
-        do! executeQueryAsync clientContext
+        let execQuery () = clientContext |> executeQuery
+        do! execQuery  ()
         return list1
     }
 
 let fixColumnName (name:string) = name.Replace(" ", "_x0020_")
-
-type SPCascadeDropDownSetup = {
-    relationshipList : string
-    relationshipListParentColumn : string
-    relationshipListChildColumn : string
-    parentColumn : string
-    childColumn : string
-    debug : bool
- }
 
 let uploadMasterPage (content) (clientContext:ClientContext) =
     let web = clientContext.get_web()
@@ -232,26 +221,33 @@ let uploadMasterPage (content) (clientContext:ClientContext) =
     
 let convert<'T> (source:Fable.Import.SharePoint.IEnumerable<'T>) : array<'T> =
     let enumerator = source.getEnumerator()
-    seq {
-        while enumerator.moveNext() do
-            yield enumerator.get_current()
-    } |> Seq.toArray
+    //logO2 "enumerator" enumerator
+    let mResult = 
+        seq {
+            while enumerator.moveNext() do
+                let result = enumerator.get_current()
+                //logO2 "result" result
+                yield result
+        } |> Seq.toArray
+    //logO2 "mResult" mResult
+    mResult
 
 [<Emit("new SP.WorkflowServices.WorkflowServicesManager($0, $1)")>]
 let WorkflowServicesManager (context:ClientContext, web:Web) = jsNative
 
 let getSubscriptionsByList (context:ClientContext) (web:Web) (listId) =   
-    async { 
+    promise { 
         let sMgr = WorkflowServicesManager (context, web)
         let sservice = sMgr?getWorkflowSubscriptionService()
         let ssubs = sservice?enumerateSubscriptionsByList(listId) :?> ClientObject
         context.load(ssubs)
-        do! executeQueryAsync context
+        let execQuery () = context |> executeQuery
+        do! execQuery  ()
         let e = ssubs?getEnumerator()
         while (e?moveNext() :?> bool) do
            let c =  e?get_current()
            log("Name :" + c?get_name().ToString() + " sID: " + c?get_id().ToString())
-    } |> Async.StartImmediate
+    } |> Promise.iter( ignore )
 
 [<Emit("SP.WorkflowServices")>]
 let WorkflowServices () = jsNative
@@ -262,26 +258,29 @@ let getSubscriptionById (context:ClientContext) (web:Web) (subscriptionId:string
     //log "Loading subscription"
     context.load(subscription)
     subscription
+    
 let startWorkFlow (context:ClientContext) (web:Web) (subscription) (itemId:int) =
-    async {
+    promise {
         let wfServiceManager = WorkflowServices()?WorkflowServicesManager?newObject(context, web)
         let inputParameters = new System.Collections.Generic.Dictionary<string, obj>()
         //log "Successfully starting workflow."
         wfServiceManager?getWorkflowInstanceService()?startWorkflowOnListItem(subscription, itemId, inputParameters) |> ignore
-        do! executeQueryAsync context
+        let execQuery () = context |> executeQuery
+        do! execQuery  ()
         //log "workflow started successfully"
     }
-    |> Async.StartImmediate
+    |> Promise.iter( ignore )
 
 let startSiteWorkFlow (context:ClientContext) (web:Web) (subscription) (inputParameters) =
-    async {
+    promise {
         let wfServiceManager = WorkflowServices()?WorkflowServicesManager?newObject(context, web)
         //log "Successfully starting workflow."
         wfServiceManager?getWorkflowInstanceService()?startWorkflow(subscription, inputParameters) |> ignore
-        do! executeQueryAsync context
+        let execQuery () = context |> executeQuery
+        do! execQuery  ()
         //log "workflow started successfully"
     }
-    |> Async.StartImmediate
+    |> Promise.iter( ignore )
 
 [<Emit("SPDragDropManager")>]
 let SPDragDropManager() = jsNative
@@ -295,15 +294,6 @@ let disableDragAndDrop () =
 
 [<Emit("ExecuteOrDelayUntilScriptLoaded($0,$1)")>]
 let ExecuteOrDelayUntilScriptLoaded (callback:unit->unit) (script:string) = jsNative
-
-let nearestFormRowParent el = 
-    try 
-        el?parents("td .ms-formbody, td .ms-formlabel")?parent()
-    with
-    | ex -> 
-        log (sprintf "nearestFormRowParent FAILED for %A [%A]" el ex )
-        null
-
 let nearestTd el = 
     el?parents("td")
 
@@ -324,13 +314,14 @@ let commonModalDialogClose(dialogResult, returnValue) = jsNative
 let rlfiShowMore () = jsNative
 
 let isCurrentUserMemberOfGroup (clientContext:ClientContext) (groupName:string) =
-    async {
+    promise {
         let web = clientContext.get_web()
         let currentUser = web.get_currentUser()
         clientContext.load(currentUser)
         let userGroups = currentUser.get_groups()
         clientContext.load(userGroups)
-        do! executeQueryAsync clientContext
+        let execQuery () = clientContext |> executeQuery
+        do! execQuery ()
 
         return userGroups |> convert<Group> |> Array.exists( fun p -> p.get_title() = groupName )
     }
@@ -339,26 +330,28 @@ let isCurrentUserMemberOfGroup (clientContext:ClientContext) (groupName:string) 
 let getSelectedItems () = jsNative
 
 let getCurrentUserAsync (clientContext:ClientContext) =
-    async {
+    promise {
         let web = clientContext.get_web()
         let currentUser = web.get_currentUser()
         clientContext.load(currentUser)
         let userGroups = currentUser.get_groups()
         clientContext.load(userGroups)
-        do! executeQueryAsync clientContext
+        let execQuery () = clientContext |> executeQuery
+        do! execQuery () 
         return currentUser
     }
 
 let getUserByIdAsync (context:ClientContext) (web: Web) (id: float) =
-   async {
+   promise {
       let user = web.get_siteUsers().getById(id)
+      let execQuery () = context |> executeQuery
       context.load(user)
-      context.executeQueryAsync()
+      do! execQuery ()
       return user
    } 
 
 let approveOrRejectTask (clientContext:ClientContext) (taskId: float) (listName: string) (approve: bool) =
-  async {
+  promise {
       let list = clientContext.get_web().get_lists().getByTitle(listName)
       let task = list.getItemById(taskId)
       let status = 
@@ -371,7 +364,8 @@ let approveOrRejectTask (clientContext:ClientContext) (taskId: float) (listName:
       task.set_item("Status",status)
       task.set_item("WorkflowOutcome",status)
       task.update()
-      do! executeQueryAsync clientContext
+      let execQuery () = clientContext |> executeQuery
+      do! execQuery ()
   }
 
 [<Emit("new SP.ClientContext($0)")>]
@@ -387,8 +381,9 @@ let getLookupFieldValue() = jsNative : FieldLookupValue
 let setFolder() = jsNative 
 
 let createFolders (context: ClientContext) (library: List) (folderNames: string []) =
-  async {
-    Browser.Support.logO ("createFolder() started")
+  promise {
+    log "createFolder() started"
+    let execQuery () = context |> executeQuery
 
     folderNames 
     |> Array.iter(fun folderName ->
@@ -403,7 +398,7 @@ let createFolders (context: ClientContext) (library: List) (folderNames: string 
             | ex -> log (sprintf "Creation of subfolder %s failed with %A" folderName ex )
     )
 
-    do! executeQueryAsync context
+    do! execQuery()
   }
 
 [<Emit("getCurrentCtx()")>]
@@ -413,7 +408,7 @@ let getCurrentCtx() = jsNative : Fable.Import.SharePoint.ContextInfo
 let getSelectedList() = jsNative 
 
 let updateColumnValue (context:ClientContext) (web: Web) (listName: string) (itemId: float) (columnName: string) (columnValue: obj) =
-  async {
+  promise {
       let item = web.get_lists().getByTitle(listName).getItemById(itemId)
       item.set_item(columnName, columnValue)
       item.update()
@@ -421,21 +416,22 @@ let updateColumnValue (context:ClientContext) (web: Web) (listName: string) (ite
   }
 
 let getAllListItems (context: ClientContext) (web: Web) (listName: string) (includePart: string) =
-  async {
+  promise {
         let list = web.get_lists().getByTitle(listName)
+        let execQuery () = context |> executeQuery
         
         match includePart.Equals("") with 
         | true -> context.load(list)
         | false -> context.load(list, includePart)
-        do! executeQueryAsync context
+        do! execQuery ()
 
         let query = CamlQuery.createAllItemsQuery()
 
         let items = list.getItems(query)
         context.load(items)
-        do! executeQueryAsync context
+        do! execQuery ()
 
-        Browser.Support.log(sprintf "listItems.get_count %A" (items.get_count()) )
+        log(sprintf "listItems.get_count %A" (items.get_count()) )
         return items
   }
 
@@ -475,14 +471,25 @@ let getVal (columnName:string) (item:ListItem)  =
 [<Emit("jQuery('<div>').html($0).text()")>]
 let convertSimpleHtmlToText(text:string) = jsNative : string
 
+let toStringSafe v =
+    if v = null then
+        ""
+    else
+        let vs = v.ToString()
+        if vs="null" then "" else vs
+
 let getValS (columnName:string) (item:ListItem)  = 
   getVal columnName item |> toStringSafe |> convertSimpleHtmlToText
 
 let getListItemsByCaml (web:Web) (context:ClientContext) (listName: string) (fieldNames: string) (camlQuery: string) (mapper)  =
-    async {
+    console.log("getListItemsByCaml started for " + listName)
+    promise {
         let itemList = web.get_lists().getByTitle(listName)
         context.load(itemList)
-        do! executeQueryAsync context
+        let execQuery () = context |> executeQuery
+        do! execQuery () 
+        //console.log(listName)
+        //console.log(itemList)
         //logO listName itemList
 
         let caml = new CamlQuery()
@@ -491,11 +498,15 @@ let getListItemsByCaml (web:Web) (context:ClientContext) (listName: string) (fie
 
         let listItems = itemList.getItems(caml)
         context.load(listItems, (sprintf "Include(Id,%s)" fieldNames ) )
-        do! executeQueryAsync context
+        do! execQuery () 
+        //console.log("listItems")
+        //console.log(listItems)
         //logO "listItems" listItems
         //logD (sprintf "listItems.get_count %A" (listItems.get_count()) )
         
         let res = listItems |> convert<ListItem> |> Array.map( mapper )
+        console.log("getListItemsByCaml succeeded for " + listName)
+        //console.log(res)
         //logO "res" res
         return res
     }